automation/auto-vpo/workflow_automation/requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=1.2.0,<2.0.0  # xlrd 2.0+ 不再支持 .xls 格式，需要使用 1.x 版本
pyyaml>=6.0.1
pywinauto>=0.6.8
pyautogui>=0.9.54
pywin32>=306
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0

